+++ test/booster_test.rb
@@ -33,6 +33,10 @@ class BoosterTest < Minitest::Test
     assert_equal Encoding::UTF_8, config.encoding
   end
 
+  def test_num_features
+    assert_equal 4, booster.num_features
+  end
+
   def test_score
     expected = {"f0" => 118, "f2" => 93, "f1" => 104, "f3" => 43}
     assert_equal expected.values.sort, booster.score.values.sort